Use large cursors with large print themes.  Fixes bug #490176.
Patch adds three new configuration options:

--with-lp-cursor-theme=THEME
--with-hclp-cursor-theme=THEME
--with-hclpi-cursor-theme=THEME

These allow you to specify which cursor theme to use with the LargePrint,
HighContrastLargePrint, and HighContrastLargePrintInverse desktop themes
by default, so that distros that ship (for example) DMZ-White and DMZ-Black
can (and probably should) specify those instead.

The defaults are whiteglass for LP and HCLPI, and redglass for HCLP, since
these are usually available on any new-ish, X.org-based desktop. Unfortunately,
their translucency and shadow makes them less-than-great choices, but the
Default cursor theme isn't resizable, so it's probably the best we can
do until it is.

The cursor size for all large print themes is set to 48px. The LargePrint
setting is also used for the LowContrastLargePrint theme, since that theme
is really so rarely-used as to be not really worth its own setting.